refactor: move class about request or response from model to request or response
