+++ src/main/java/bitnagil/bitnagil_backend/auth/oauth2/response/KakaoUserInfo.java
@@ -2,7 +2,6 @@ package bitnagil.bitnagil_backend.auth.oauth2.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import bitnagil.bitnagil_backend.auth.oauth2.model.KakaoAccount;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
