+++ src/main/java/bitnagil/bitnagil_backend/auth/oauth2/service/AuthService.java
@@ -11,13 +11,12 @@ import bitnagil.bitnagil_backend.auth.jwt.JwtProvider;
 import bitnagil.bitnagil_backend.auth.oauth2.response.KakaoTokenResponse;
 import bitnagil.bitnagil_backend.global.errorcode.ErrorCode;
 import bitnagil.bitnagil_backend.global.exception.CustomException;
-import bitnagil.bitnagil_backend.global.response.CustomResponseDto;
 import bitnagil.bitnagil_backend.user.Repository.UserRepository;
 import bitnagil.bitnagil_backend.enums.SocialType;
 import bitnagil.bitnagil_backend.auth.jwt.TokenResponse;
 import bitnagil.bitnagil_backend.user.entity.User;
 import bitnagil.bitnagil_backend.enums.Role;
-import bitnagil.bitnagil_backend.auth.oauth2.model.KakaoAccount;
+import bitnagil.bitnagil_backend.auth.oauth2.response.KakaoAccount;
 import bitnagil.bitnagil_backend.auth.oauth2.response.KakaoUserInfo;
 import lombok.RequiredArgsConstructor;
